pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.version]
path = "src/proj_flow/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs"]

[tool.hatch.build.targets.wheel]
packages = ["src/proj_flow"]

[project]
authors = [{name = "Marcin Zdun", email = "marcin.zdun@gmail.com"}]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
  "argcomplete",
  "chevron2021",
  "prompt_toolkit",
  "PyYAML",
]
description = "C++ project maintenance, automated"
dynamic = ["version"]
keywords = ['C/C++', 'developer', 'manager', 'dependency', 'tool', 'ci', 'c++', 'cpp']
name = "proj-flow"
readme = "README.md"
requires-python = ">=3.10"

[project.urls]
Documentation = "https://proj-flow.readthedocs.io/en/latest/"
Homepage = "https://pypi.org/project/proj-flow/"
"Source Code" = "https://github.com/mzdun/proj-flow"

[project.scripts]
cxx-flow = "proj_flow.flow.cli:main"
proj-flow = "proj_flow.cli:main"
